Allow default_code to use earlier field values
For required fields, you could already access the value as `self.foo.0`,
and for other fields by abusing implementation details, but that was all
a really nasty way of doing it. This is much nicer.

As noted in the changelog, this is a breaking change, but extremely
unlikely to actually break any code, in part due to case conventions.
Here’s a simple demonstration of the sort of code that it would break
(in this case silently, but compilation failure is also possible if the
shadowed value is of a different type):

    pub const x: i32 = 1;

    struct Foo {
	x: i32,
	#[builder(default_code = "x")]
	y: i32,
    }

    Foo::builder().x(2).build().y

Before this commit, y will be 1, from the x const.
After this commit, y will be 2, from the x field.

One mildly discomfiting aspect of this feature is that it allows the
order of field declaration to be significant. As noted in the code, we
can fix this later with a field dependency DAG guiding reordering. That
would also be a technically breaking change for the same reasons as this
commit is, but given this feature you’d have to be somewhat insane to
write code that would be broken by it!

A reasonable extension of this work is allowing the user to add prelude
statements to the build() method, so values to be used by the default
calculations of multiple methods can be defined.

An alternative I considered to avoid this being a breaking change (if we
consider current use of `self` to be not public API, and thus fair game
to break) is giving the locals namespaced names so that `foo` would be
`TypeBuilder_field_foo`, transforming `default_code` so that `self.foo`
would become `TypedBuilder_field_foo`. But that’s harder, and an abuse
of `self` anyway. Better, I think, to do it this way.