Better support documentation
This doesn’t actually solve #11 (there are still substantial changes
that should be made to the default docs for the `builder()` method), but
it lets you customise how it will appear, and unhide the builder type,
which are both a Good Thing™.

I have written no real tests for this because the system’s not
well-placed for doing that; I’ve just written tests to ensure that it
at least compiles. Yet not testing the documentation generated is not
good.