Rename TypedBuilder_BuilderFor_Foo to FooBuilder
I expect this to cause no trouble at all, and it’s a much nicer default.
I see no reason for “TypedBuilder” to appear anywhere in the generated
code or docs.

`#[builder(name = TypedBuilder_BuilderFor_Foo)]` will restore the old
behaviour if anyone really needs or wants it for some bizarre reason.

I tossed up between `#[builder(name = "StringLiteral")]` and
`#[builder(name = Identifier)]` forms; historically, attributes with `=`
in Rust were all `key = "value"`, and identifier values were only used
in the `function(value)` style (e.g. with `derive(…)` and `cfg(…)`).
(This was actually a syntactic rule; the meta matcher used to be much
tighter, but was relaxed a lot somewhere along the way, not long before
Rust 1.0 if I recall correctly.) Although this is still the general
convention, it’s weaker than it used to be, and this is a case where I
think breaking the convention is probably warranted, as the value will
be used as an identifier in the generated code.

In the end, Identifier was what I implemented, completely by accident (I
confess I thought I’d implemented the other), so I stuck with it.